Added minor features
Made the navbar stick to the top of the page. It will now persist and
follow the user’s scrolls. Only implemented this for map.html and
dataEntry.html. Still working on adding it to index.html. I also
improved the positioning of the status bar. It is now consistent across
the map.html and dataEntry.html pages. Still need to improve the its
positioning on index.html.

Also added some comments to maincss.css and index.html.